Attempt to have the code promisified
This code does not work as I expected.
The problem is that the promise.then callback
is NOT executed while this iteration is running.
It appeared to be run and the next outer forEach loop
when a new request is done.
As a consequence the time captured will be only captured when
all implementations in this forEach loop have executed.

There is an excellent article
https://jakearchibald.com/2015/tasks-microtasks-queues-and-schedules/
about which timing can be expected.

However it appears that the triggering I observed with
Chrome Version 47.0.2526.106 (64-bit) on OS X 10.10.5
may not be explained there.